+++ packages/@dcl/inspector/src/components/Toolbar/Button/Button.css
@@ -1,0 +1,29 @@
+.Button {
+  border: none;
+  padding: 0px 2px;
+  margin: 0 8px;
+  color: #fff;
+  background-color: #4b4852;
+  font-size: 16px;
+  border-radius: 4px;
+  cursor: pointer;
+  min-width: 32px;
+}
+
+.Button:hover:not(:disabled) {
+  background-color: #6c6974;
+}
+
+.Button:disabled {
+  opacity: 0.5;
+  cursor: default
+}
+
+.Button.active {
+  background-color: #6c6974;
+}
+
+.Button svg {
+  width: 24px;
+  height: 24px;
+}
