+++ packages/@dcl/inspector/src/components/Toolbar/Gizmos/index.ts
@@ -1,0 +1,2 @@
+import Gizmos from './Gizmos'
+export { Gizmos }
